requirements.txt
requests>=2.28.0
flask>=2.2.0
python-dotenv>=0.19.0
Pillow>=9.0.0
PyYAML>=6.0.0
